:test_tube: [#94] Add tests
