+++ backend/src/openarchiefbeheer/destruction/tests/test_endpoints.py
@@ -1,4 +1,3 @@
-from datetime import date
 from unittest.mock import patch
 
 from django.core import mail
@@ -678,146 +677,6 @@ class DestructionListViewSetTest(APITestCase):
             response.json()["user"][0],
             _("The chosen user does not have the permission to review a final list."),
         )
-
-    def test_start_destruction(self):
-        record_manager = UserFactory.create(
-            username="record_manager", role__can_start_destruction=True
-        )
-        destruction_list = DestructionListFactory.create(
-            name="A test list",
-            contains_sensitive_info=True,
-            author=record_manager,
-            status=ListStatus.ready_to_delete,
-        )
-
-        self.client.force_authenticate(user=record_manager)
-        with patch(
-            "openarchiefbeheer.destruction.api.viewsets.delete_destruction_list"
-        ) as m_delete:
-            response = self.client.delete(
-                reverse(
-                    "api:destructionlist-detail", kwargs={"uuid": destruction_list.uuid}
-                ),
-            )
-
-        self.assertEqual(status.HTTP_204_NO_CONTENT, response.status_code)
-        m_delete.assert_called_once()
-        self.assertEqual(m_delete.call_args_list[0].args[0].pk, destruction_list.pk)
-
-        destruction_list.refresh_from_db()
-
-        self.assertEqual(destruction_list.processing_status, InternalStatus.queued)
-
-    def test_cannot_start_destruction_if_not_author(self):
-        record_manager = UserFactory.create(role__can_start_destruction=True)
-        destruction_list = DestructionListFactory.create(
-            name="A test list",
-            contains_sensitive_info=True,
-            status=ListStatus.ready_to_delete,
-        )
-
-        self.client.force_authenticate(user=record_manager)
-        with patch(
-            "openarchiefbeheer.destruction.api.viewsets.delete_destruction_list"
-        ) as m_task:
-            response = self.client.delete(
-                reverse(
-                    "api:destructionlist-detail", kwargs={"uuid": destruction_list.uuid}
-                ),
-            )
-
-        self.assertEqual(status.HTTP_403_FORBIDDEN, response.status_code)
-        m_task.assert_not_called()
-
-    def test_cannot_start_destruction_if_not_ready_to_delete(self):
-        record_manager = UserFactory.create(role__can_start_destruction=True)
-        destruction_list = DestructionListFactory.create(
-            name="A test list",
-            contains_sensitive_info=True,
-            author=record_manager,
-            status=ListStatus.ready_to_review,
-        )
-
-        self.client.force_authenticate(user=record_manager)
-        with patch(
-            "openarchiefbeheer.destruction.api.viewsets.delete_destruction_list"
-        ) as m_task:
-            response = self.client.delete(
-                reverse(
-                    "api:destructionlist-detail", kwargs={"uuid": destruction_list.uuid}
-                ),
-            )
-
-        self.assertEqual(status.HTTP_403_FORBIDDEN, response.status_code)
-        m_task.assert_not_called()
-
-    def test_cannot_start_destruction_if_archiefactiedatum_in_the_future(self):
-        record_manager = UserFactory.create(
-            username="record_manager", role__can_start_destruction=True
-        )
-        destruction_list = DestructionListFactory.create(
-            name="A test list",
-            author=record_manager,
-            status=ListStatus.ready_to_delete,
-        )
-        DestructionListItemFactory.create(
-            with_zaak=True,
-            zaak__archiefactiedatum=date(2025, 1, 1),
-            destruction_list=destruction_list,
-            status=ListItemStatus.suggested,
-        )
-        DestructionListItemFactory.create(
-            with_zaak=True,
-            zaak__archiefactiedatum=date(2023, 1, 1),
-            destruction_list=destruction_list,
-            status=ListItemStatus.suggested,
-        )
-
-        self.client.force_authenticate(user=record_manager)
-        with freezegun.freeze_time("2024-01-01T21:36:00+02:00"):
-            response = self.client.delete(
-                reverse(
-                    "api:destructionlist-detail", kwargs={"uuid": destruction_list.uuid}
-                ),
-            )
-
-        self.assertEqual(status.HTTP_400_BAD_REQUEST, response.status_code)
-        self.assertEqual(
-            response.json()[0],
-            _("This list contains cases with archiving date in the future."),
-        )
-
-    def test_can_start_destruction_if_archiefactiedatum_in_the_future_but_removed(self):
-        record_manager = UserFactory.create(
-            username="record_manager", role__can_start_destruction=True
-        )
-        destruction_list = DestructionListFactory.create(
-            name="A test list",
-            author=record_manager,
-            status=ListStatus.ready_to_delete,
-        )
-        DestructionListItemFactory.create(
-            with_zaak=True,
-            zaak__archiefactiedatum=date(2025, 1, 1),
-            destruction_list=destruction_list,
-            status=ListItemStatus.removed,
-        )
-        DestructionListItemFactory.create(
-            with_zaak=True,
-            zaak__archiefactiedatum=date(2023, 1, 1),
-            destruction_list=destruction_list,
-            status=ListItemStatus.suggested,
-        )
-
-        self.client.force_authenticate(user=record_manager)
-        with freezegun.freeze_time("2024-01-01T21:36:00+02:00"):
-            response = self.client.delete(
-                reverse(
-                    "api:destructionlist-detail", kwargs={"uuid": destruction_list.uuid}
-                ),
-            )
-
-        self.assertEqual(status.HTTP_204_NO_CONTENT, response.status_code)
 
     def test_mark_as_ready_to_review(self):
         record_manager = UserFactory.create(role__can_start_destruction=True)
